backend/requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-Cors==4.0.0
Werkzeug==2.3.7

# Environment and configuration
python-dotenv==1.0.0

# API integrations
twilio==8.9.1
openai==0.28.1
requests==2.31.0

# File handling
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.1
openpyxl==3.1.2

# NLP and embeddings
scikit-learn==1.3.1
nltk==3.8.1
sentence-transformers==2.2.2

# Database
SQLAlchemy==2.0.21

# Utilities
uuid==1.30
python-dateutil==2.8.2
pytz==2023.3.post1